% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricanes.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(raw.data, ...)
}
\arguments{
\item{raw.data}{tibble output of \code{load_data()}}
}
\value{
This function returns a tibble that contains the columns \code{storm_id},
\code{date},\code{latitude}, \code{longitude}, \code{wind_speed}, \code{ne},
\code{nw}, \code{se}, \code{sw}
}
\description{
This is a function that takes the output of \code{load_data()} (above) in a 
tibble format, creates a new column called \code{storm_id} where \code{storm_name} 
and \code{year} are merged with a hyphen (ex: "KATRINA-2005"). A second column is
also created, \code{date}, that reformats the month, day, year, and hour data from
the original data to have one ymd_hms formatted column. The data is then gathered
by the 4 directions x 3 windspeeds. The key of this is then separated to retrieve
\code{windspeed}, \code{radius}, and \code{direction}. The \code{direction} is then
spread into 4 columns, one for each direction. The \code{longitude} is mutated to
be negative and account for the western hemisphere. Finally, specific columns
are selected for the output tibble.
}
\examples{
clean_data(raw.data)

}
