% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricanes.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Geom Function for Hurricane Data}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\value{
Geom Hurricane layer
}
\description{
This function uses GeomHurricane to create a geom function to be used
to layer the GeomHurricane image onto a base plot or map
}
\examples{
map_data <- get_map("Louisiana", zoom = 6, maptype = "toner-background")
base_map <- ggmap(map_data, extent = "device")

base_map +
  geom_hurricane(data = hurricane, aes(x = longitude, y = latitude,
                                       r_ne = ne, r_se = se,
                                       r_nw = nw, r_sw = sw,
                                       fill = wind_speed,
                                       color = wind_speed))

}
