% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricanes.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{ggproto function for hurricane}
\format{An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomHurricane
}
\value{
Geom including the image to be lyared
}
\description{
This is the creation of a ggproto object to create 4 quadrants of colored
images of the NE, NW, SE, and SW winds.
}
\examples{
map_data <- get_map("Louisiana", zoom = 6, maptype = "toner-background")
base_map <- ggmap(map_data, extent = "device")

base_map +
  geom_hurricane(data = hurricane, aes(x = longitude, y = latitude,
                                       r_ne = ne, r_se = se,
                                       r_nw = nw, r_sw = sw,
                                       fill = wind_speed,
                                       color = wind_speed))

}
\keyword{datasets}
