% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricanes.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(filename = "ebtrk_atlc_1988_2015.txt", ...)
}
\arguments{
\item{filename}{A character string giving the path and name of the file
to be read in. Default value is "ebtrk_atlc_1988_2015.txt"}
}
\value{
This function returns a tibble that contains the columns in the 
read in file
}
\description{
This is a simple function that loads hurricane data from 
http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/. 
Using column widths and assigned colnames, the data is read in using the 
\code{readr::read_fwf}.
}
\examples{
load_data()
load_data(filename = "ebtrk_atlc_1988_2015.txt")

}
