% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricanes.R
\name{select_hurricane}
\alias{select_hurricane}
\title{Select Hurricane Data}
\usage{
select_hurricane(data, storm_date = "KATRINA-2005",
  date_time = "2005290812", ...)
}
\arguments{
\item{data}{tibble output of \code{clean_data()}}
}
\value{
This function returns a tibble contains \code{storm_name} and 
\code{date} such that the output is specifically related to one date and time.
}
\description{
This is a function that takes the output of \code{clean_data()} (above) in a 
tibble format and filters by \code{storm_name} and \code{date} such that the 
output is specifically related to one date and time.
}
\examples{
select_hurricane(data)
select_hurricane(data, storm_date='KATRINA-2005', date_time = '2005290812')
}
